+++ lime-runtime/src/main/java/com/here/gluecodium/model/lime/LimeConstant.kt
@@ -21,9 +21,8 @@ package com.here.gluecodium.model.lime
 
 class LimeConstant(
     path: LimePath,
-    visibility: LimeVisibility = LimeVisibility.PUBLIC,
     comment: LimeComment = LimeComment(),
     attributes: LimeAttributes? = null,
     typeRef: LimeTypeRef,
     val value: LimeValue
-) : LimeTypedElement(path, visibility, comment, attributes, typeRef = typeRef)
+) : LimeTypedElement(path, comment, attributes, typeRef = typeRef)