+++ lime-runtime/src/main/java/com/here/gluecodium/model/lime/LimeInterface.kt
@@ -21,7 +21,6 @@ package com.here.gluecodium.model.lime
 
 class LimeInterface(
     path: LimePath,
-    visibility: LimeVisibility = LimeVisibility.PUBLIC,
     comment: LimeComment = LimeComment(),
     attributes: LimeAttributes? = null,
     external: LimeExternalDescriptor? = null,
@@ -39,7 +38,6 @@ class LimeInterface(
     val isNarrow: Boolean = false
 ) : LimeContainerWithInheritance(
     path = path,
-    visibility = visibility,
     comment = comment,
     attributes = attributes,
     external = external,